seperate datatree into its own file
